Tipo de busca por ID de Long para Integer no metodo buscarPorId()

+++ Sistema/002-IMPLEMENTACAO/SISAP/src/main/java/br/edu/ifpb/monteiro/ads/sisap/dao/AtendimentoDAO.java
@@ -42,7 +42,7 @@ public class AtendimentoDAO extends DAO {
 		return resultado;
 	}
 
-	public Atendimento buscarPorId(Long id) throws SisapException {
+	public Atendimento buscarPorId(Integer id) throws SisapException {
 		EntityManager em = getEntityManager();
 		Atendimento resultado = null;
 		try {
